# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  dto.ReviewListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ReviewResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.ReviewRequest:
    properties:
      comment:
        maxLength: 500
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - rating
    type: object
  dto.ReviewResponse:
    properties:
      author_display:
        type: string
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      rating:
        type: integer
    type: object
  dto.ReviewSummary:
    properties:
      average:
        type: number
      count:
        type: integer
      distribution:
        additionalProperties:
          type: integer
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/leoferamos/aroma-sense
  description: REST API for user Aroma Sense Eccomerce
  title: Aroma Sense API
  version: "1.0"
paths:
...
  /products/{id}/reviews:
    get:
      consumes:
      - application/json
      description: Returns published reviews for a product in descending creation
        order.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated reviews
          schema:
            $ref: '#/definitions/dto.ReviewListResponse'
        "400":
          description: Invalid product id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List product reviews
      tags:
      - reviews
    post:
      consumes:
      - application/json
      description: Creates a review for a delivered product. Requires authentication,
        a display_name set, and at least one delivered order containing the product.
        One review per user/product.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Review payload
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/dto.ReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Review created
          schema:
            $ref: '#/definitions/dto.ReviewResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden (not delivered or profile incomplete)
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Already reviewed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create product review
      tags:
      - reviews
  /products/{id}/reviews/summary:
    get:
      consumes:
      - application/json
      description: Returns average rating, total review count, and rating distribution
        for a product.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Review summary
          schema:
            $ref: '#/definitions/dto.ReviewSummary'
        "400":
          description: Invalid product id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get product review summary
      tags:
      - reviews
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"